+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- `@Bound` annotation (works like `@BoundTo`, but implicitly uses the direct supertype of the annotated class)
+
 ### Changed
+- Make `component` in annotations optional and set to `SingletonComponent` by default.
 - Dependencies:
   - Upgrade Hilt to `2.31-alpha`
   - Upgrade `org.jetbrains.changelog` to `1.0.1`